+++ extensions/perplexity-api/src/hook/utils.ts
@@ -5,11 +5,13 @@ export const allModels = [
   { name: "Mixtral 8x7b Instruct", id: "mixtral-8x7b-instruct" },
   { name: "Mistral 7B Instruct", id: "mistral-7b-instruct" },
   { name: "Llama2 70B Chat", id: "llama-2-70b-chat" },
-  { name: "Perplexity 7B Chat", id: "pplx-7b-chat" },
+  { name: "Sonar small 7B Chat", id: "sonar-small-chat" },
+  { name: "Sonar medium 8x7B Chat", id: "sonar-medium-chat" },
   { name: "Perplexity 70B Chat", id: "pplx-70b-chat" },
   { name: "CodeLlama 34B Instruct", id: "codellama-34b-instruct" },
   { name: "CodeLlama 70B Instruct", id: "codellama-70b-instruct" },
-  { name: "Perplexity 7B Online", id: "pplx-7b-online" },
+  { name: "Sonar small 7B Online", id: "sonar-small-online" },
+  { name: "Sonar medium 8x7B Online", id: "sonar-medium-online" },
   { name: "Perplexity 70B Online", id: "pplx-70b-online" },
 ];
 
@@ -32,12 +34,12 @@ export function countToken(content: string) {
 export function estimatePrice(prompt_token: number, output_token: number, model: string) {
   let price = 0;
   switch (model) {
-    case "pplx-7b-chat":
+    case "sonar-small-chat":
     case "mistral-7b-instruct":
       price = ((prompt_token * 0.07) / 1_000_000 + (output_token * 0.28) / 1_000_000) * 100;
       break;
     case "mixtral-8x7b-instruct":
-    case "pplx-8x7b-chat":
+    case "sonar-medium-chat":
       price = ((prompt_token * 0.6) / 1_000_000 + (output_token * 1.8) / 1_000_000) * 100;
       break;
     case "codellama-34b-instruct":
@@ -48,10 +50,10 @@ export function estimatePrice(prompt_token: number, output_token: number, model:
     case "pplx-70b-chat":
       price = ((prompt_token * 0.7) / 1_000_000 + (output_token * 2.8) / 1_000_000) * 100;
       break;
-    case "pplx-7b-online":
+    case "sonar-small-online":
       price = (5 / 1000 + (output_token * 0.28) / 1_000_000) * 100;
       break;
-    case "pplx-8x7b-online":
+    case "sonar-medium-online":
       price = (5 / 1000 + (output_token * 1.8) / 1_000_000) * 100;
       break;
     case "pplx-70b-online":